+++ Entities/RunStatsEntity.cs
@@ -5,10 +5,10 @@ namespace SurfTimer.Shared.Entities
     public abstract class RunStatsEntity
     {
         public int ID { get; set; }
-        public int Type { get; set; }      // 0 = Map, 1 = Bonus, 2 = Stage
+        public short Type { get; set; }      // 0 = Map, 1 = Bonus, 2 = Stage
         public int RunTime { get; set; }
-        public int Stage { get; set; }     // Signifies Bonus number when Type == 1
-        public int Style { get; set; }
+        public short Stage { get; set; }     // Signifies Bonus number when Type == 1
+        public short Style { get; set; }
         public string? Name { get; set; }
         public float StartVelX { get; set; }
         public float StartVelY { get; set; }
